PROJ1716: Fix the type used to send enum in autogen code
The autogenerated code in the API was treating enums as uint8 instead of
uint32 (which is what is sent through the lower layers).
This worked for individual enums, but resulted in not enough memory being
allocated for repeated enums. This caused a crash.

The fix is to change the type used in the autogenerated code from uint8 to
uint32.

+++ src/google/protobuf/compiler/c/c_atl_generator.cc
@@ -496,7 +496,7 @@ string AtlCodeGenerator::TypeToString(FieldDescriptor::Type type)
     	description = "uint32_t";
     	break;
     case FieldDescriptor::TYPE_ENUM:
-    	description = "uint8_t";
+    	description = "uint32_t";
     	break;
     case FieldDescriptor::TYPE_SFIXED32:
     	description = "int32_t";
